+++ start/src/router/index.ts
@@ -8,7 +8,86 @@
 
 // export default router
 
-export const accessRoutes = [{ 'path': '/app/chainBrowser', 'redirect': '/app/chainBrowser/index', 'name': 'explorer', 'meta': { 'title': '浏览器', 'icon': 'left_laq', 'prefix': 'icon-', 'auth': ['explorer'] }, 'props': { 'default': false }, 'instances': {}, 'leaveGuards': {}, 'updateGuards': {}, 'enterCallbacks': {}, 'components': { 'default': {} } }, { 'path': '/app/node', 'redirect': '/app/node/nodeManage', 'name': 'node', 'meta': { 'title': '节点管理', 'keepAlive': true, 'icon': 'left_jd', 'prefix': 'icon-' }, 'props': { 'default': false }, 'children': [{ 'path': '/app/node/nodeManage', 'name': 'node_nodeManage', 'meta': { 'title': '节点列表', 'auth': ['node-list'] } }, { 'path': '/app/node/votingManage', 'name': 'node_votingManage', 'meta': { 'title': '节点入网投票', 'auth': ['node-vote'] } }, { 'path': '/app/node/consensusManage', 'name': 'node_consensusManage', 'meta': { 'title': '节点共识投票', 'auth': ['node-consensus'] } }], 'instances': {}, 'leaveGuards': {}, 'updateGuards': {}, 'enterCallbacks': {}, 'components': { 'default': {} } }, { 'path': '/app/contract', 'redirect': '/app/contract/index', 'name': 'contract', 'meta': { 'title': '合约管理', 'keepAlive': true, 'icon': 'left_jd', 'prefix': 'icon-' }, 'props': { 'default': false }, 'children': [{ 'path': '/app/contract/index', 'name': 'contract_index', 'meta': { 'title': '合约列表' } }], 'instances': {}, 'leaveGuards': {}, 'updateGuards': {}, 'enterCallbacks': {}, 'components': { 'default': {} } }, { 'path': '/app/cross-chain-routing', 'redirect': '/app/cross-chain-routing/index', 'name': 'cross-chain-routing', 'meta': { 'title': '跨链路由', 'keepAlive': true, 'icon': 'left_jd', 'prefix': 'icon-' }, 'props': { 'default': false }, 'children': [{ 'path': '/app/cross-chain-routing/index', 'name': 'cross-chain-routing_index', 'meta': { 'title': '路由信息' } }], 'instances': {}, 'leaveGuards': {}, 'updateGuards': {}, 'enterCallbacks': {}, 'components': { 'default': {} } }, { 'path': '/app/among/data', 'redirect': '/app/among/data/index', 'name': 'monitor', 'meta': { 'title': '数链管理', 'keepAlive': true, 'icon': 'left_jd', 'prefix': 'icon-' }, 'props': { 'default': false }, 'children': [{ 'path': '/app/among/data/index', 'name': 'monitor_index', 'meta': { 'title': '上链数据查询' } }], 'instances': {}, 'leaveGuards': {}, 'updateGuards': {}, 'enterCallbacks': {}, 'components': { 'default': {} } }]
+export const accessRoutes = [
+  {
+    path: '/app/chainBrowser',
+    redirect: '/app/chainBrowser/index',
+    name: 'explorer',
+    meta: { title: '浏览器', icon: 'left_laq', prefix: 'icon-', auth: ['explorer'] },
+    props: { default: false },
+    instances: {},
+    leaveGuards: {},
+    updateGuards: {},
+    enterCallbacks: {},
+    components: { default: {} }
+  },
+  {
+    path: '/app/node',
+    redirect: '/app/node/nodeManage',
+    name: 'node',
+    meta: { title: '节点管理', keepAlive: true, icon: 'left_jd', prefix: 'icon-' },
+    props: { default: false },
+    children: [
+      { path: '/app/node/nodeManage', name: 'node_nodeManage', meta: { title: '节点列表', auth: ['node-list'] } },
+      {
+        path: '/app/node/votingManage',
+        name: 'node_votingManage',
+        meta: { title: '节点入网投票', auth: ['node-vote'] }
+      },
+      {
+        path: '/app/node/consensusManage',
+        name: 'node_consensusManage',
+        meta: { title: '节点共识投票', auth: ['node-consensus'] }
+      }
+    ],
+    instances: {},
+    leaveGuards: {},
+    updateGuards: {},
+    enterCallbacks: {},
+    components: { default: {} }
+  },
+  {
+    path: '/app/contract',
+    redirect: '/app/contract/index',
+    name: 'contract',
+    meta: { title: '合约管理', keepAlive: true, icon: 'left_jd', prefix: 'icon-' },
+    props: { default: false },
+    children: [{ path: '/app/contract/index', name: 'contract_index', meta: { title: '合约列表' } }],
+    instances: {},
+    leaveGuards: {},
+    updateGuards: {},
+    enterCallbacks: {},
+    components: { default: {} }
+  },
+  {
+    path: '/app/cross-chain-routing',
+    redirect: '/app/cross-chain-routing/index',
+    name: 'cross-chain-routing',
+    meta: { title: '跨链路由', keepAlive: true, icon: 'left_jd', prefix: 'icon-' },
+    props: { default: false },
+    children: [
+      { path: '/app/cross-chain-routing/index', name: 'cross-chain-routing_index', meta: { title: '路由信息' } }
+    ],
+    instances: {},
+    leaveGuards: {},
+    updateGuards: {},
+    enterCallbacks: {},
+    components: { default: {} }
+  },
+  {
+    path: '/app/among/data',
+    redirect: '/app/among/data/index',
+    name: 'monitor',
+    meta: { title: '数链管理', keepAlive: true, icon: 'left_jd', prefix: 'icon-' },
+    props: { default: false },
+    children: [{ path: '/app/among/data/index', name: 'monitor_index', meta: { title: '上链数据查询' } }],
+    instances: {},
+    leaveGuards: {},
+    updateGuards: {},
+    enterCallbacks: {},
+    components: { default: {} }
+  }
+]
 
 // [
 //   {